reading from stdin, reorganized
